slackbot/requirements.txt
python-dotenv==1.0.0
slack-bolt==1.18.1
aiohttp==3.9.1
peewee==3.17.0
prettytable==3.9.0
discord.py==2.3.2
peewee-migrate==1.12.2
